Multiple fixes
The main update is that the confirmation is now by E not by H. In addition, it is required that various numbers of true propositions are represented in the 100 runs per number of possible worlds (e.g., 0 true propositions, 1 true of n, 2 of n, etc). The results also inlude n=2 cases now.